requirements.txt
matplotlib
numpy
tqdm
numpy-stl
